fake sum return dot product real result

+++ sum.h
@@ -2,4 +2,4 @@
 
 int gcc_sum(float *r, float *x, float *y, unsigned int n);
 
-int fake_gcc_sum(float *r, float *x, float *y, unsigned int n);
+double fake_gcc_sum(float *r, float *x, float *y, unsigned int n);
